pyproject.toml
[tool.poetry]
name = "blog-backend-gpt"
version = "0.1.0"
description = ""
authors = ["Linghua Zhang <lz101@rice.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.*"
fastapi = {extras = ["standard"], version = "^0.115.6"}
loguru = "^0.7.3"
sqlalchemy = { version = "^2.0.21", extras = ["mypy", "asyncio"] }
yarl = "^1.18.3"
pydantic-settings = "^2.7.1"
uvicorn = "^0.34.0"
aiomysql = "^0.2.0"
tiktoken = "^0.8.0"
ujson = "^5.10.0"
langchain = "^0.3.14"
langchain-community = "^0.3.14"
replicate = "^1.0.4"
greenlet = "^3.1.1"
cryptography = "^44.0.0"
google-search-results = "^2.4.2"
langchain-text-splitters = "^0.3.5"
langgraph = "^0.2.67"
langchain-openai = "^0.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
